Create Entities namespace
