+++ example.rb
@@ -2,7 +2,7 @@ require "open_meteo"
 
 client = OpenMeteo::Client.new(agent: Faraday.new { |conn| conn.response :logger })
 
-location = OpenMeteo::Location.new(longitude: 52.52, latitude: 13.41)
+location = OpenMeteo::Entities::Location.new(longitude: 52.52, latitude: 13.41)
 variables = { current: %i[weathercode], hourly: %i[], daily: %i[] }
 forecast_response = OpenMeteo::Forecast.new(client:).get(location:, variables:)
